술 추천페이지 구현중 #54

+++ app/src/main/java/com/example/a_level/keyword/PreferenceModel.kt
@@ -3,7 +3,7 @@ package com.example.a_level.keyword
 import com.google.gson.annotations.SerializedName
 
 data class PreferenceRequest(
-    @SerializedName("user_id") val user_id: Int,
+    @SerializedName("user_id") val user_id: Long,
     @SerializedName("type") val type: String,
     @SerializedName("volumne") val volume: String,
     @SerializedName("sugar") val sugar:Int,
